// src/types/dataBank.ts
export interface IDataBank {
    id_data_offices: number
    sale_point_name: string
    address: string
    latitude: string
    longitude: string
    workload: number
    bank_monday_time: string
    bank_tuesday_time: string
    bank_thursday_time: string
    bank_friday_time: string
    bank_saturday_time: string
    bank_sunday_time: string
    terminal_monday_time: string
    terminal_tuesday_time: string
    terminal_thursday_time: string
    terminal_friday_time: string
    terminal_saturday_time: string
    terminal_sunday_time: string
    id_service: number
}